+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 15.2.1
+## 15.3.0
 
 - REVERT change in 15.2.0. `keydown` should indeed select options. This is how a native select box works.
 - Allow `Space` to select an option similar to `Enter`